Command tester no longer gets confused given a particular xmin
When testing a plot command, some arbitrary data is sent to gnuplot. If the user
asked for particular plot bounds (xmin, etc) then it was possible for the test
plot to fail if the arbitrary data was out of bounds. I now have a whitelist of
errors that are allowed during plot testing, and this error is now on that list.

Furthermore, because I have this whitelist, I removed the extra logic that was a
part of the tests, as it's unnecessary. Previously tests sent

 plot whatever; print 'xxxxxxx';

The print only runs if gnuplot thought the plot succeeded, so I was looking for
the print output. I no longer need this. I simply run the plot, clear out any
known errors, and barf if anything remains.