Check that cbrange isn't being influenced by hidden BAD values.
When doing a simple plot of an image that has some BAD values (in
my case an SDO/HMI magnetogram), using
$w->plot({with=>'image'},$im);
resulted in a correct cbrange of [minvalue, maxvalue], but using
$w->plot({with=>'fits'},$im);
resulted in a cbrange that was [-1.79769..E+308, maxvalue].
This is because (I think) the fits prefrobnicator does a
PDL::Transform::map on $im, which does not handle BAD values, so they
get set to the bad value but returned as a floating point.

This patch just checks if there are any elements with that type's
badvalue and if so, makes sure they are set BAD so the isfinite
line can handle them properly.