Set XDG_RUNTIME_DIR variable to temporary directory
If this environment variable is not set, some Gnuplot terminals (Qt)
will fail. This bug was reported by Ingo Schmid on the pdl-devel mailing
list by running `sudo cpanm PDL::Graphics::Simple` which failed some
tests with the error:

> Gnuplot error: "QStandardPaths: XDG_RUNTIME_DIR not set, defaulting to
> '/tmp/runtime-root'" while sending final multiplot command

To work around this, locally set the variable to a temporary directory
before opening the Gnuplot process.